testing opencv with pyside
