+++ logfile.txt
@@ -1,0 +1,1 @@
+09-Jul-23 17:22:01 - Start PYthon GUI App
